pdf-sidecar/requirements.txt
# PDF Sidecar Dependencies
PyMuPDF>=1.24.0
Pillow>=11.0.0

# Development dependencies (included in PyInstaller bundle)
